处理好友请求完成 查询我的好友正在做

+++ user-service-app-8001/src/main/java/com/mr/mapper/UserFriendRequestMapper.java
@@ -22,5 +22,5 @@ public interface UserFriendRequestMapper {
 
     int updateByPrimaryKey(UserFriendRequest record);
 
-    int updateStatusByPrimaryKey(@Param("id") Long id, @Param("status") Long status);
+    int updateStatusByPrimaryKey(@Param("id") Long id, @Param("status") Integer status);
 }
